image_tag/requirements.txt
torch
transformers
keybert
sentence-transformers
Pillow
faiss-cpu
numpy
flask-cors